pyproject.toml
[project]
name = "remote-command"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.6"
dependencies = [
    "ansible>=8.7.0",
    "ansible-runner>=2.4.1",
    "django>=4.2.23",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "pydantic>=2.11.7",
]
